Improve CaseRunner storage
Up to now, we stored the runners statically withing the
CaseRegistrar. This is convenient and fast to compile, but
incorrect since the same runner type can only ever be registered
once.

To fix this, we rather store the runner within the cases directly.
This changes the runner from being a std::function to a
AbstractCaseRunner. This setup is a bit more complicated, but
supports move-only runners right from the get go.

Also add a more complicated test that would have detected the
incorrectly stored runners in the first place.